% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stepCAT.choice}
\alias{stepCAT.choice}
\alias{stepCAT.choice.concerto}
\alias{stepCAT.choice.console}
\alias{stepCAT.choice.tcltk}
\title{Shows the choices for the question and reads a response}
\usage{
stepCAT.choice(GUI = "console", item = NULL, question = NA,
  choices = rep(NA, 5), ...)

\method{stepCAT.choice}{console}(item = NULL, question = NA,
  choices = rep(NA, 5), ...)

\method{stepCAT.choice}{concerto}(item = NULL, question = NA,
  choices = rep(NA, 5), templateID = NULL, templateWorkspaceID = NULL,
  submit = "Submit", console.force = FALSE, ...)

\method{stepCAT.choice}{tcltk}(item = NULL, question = NA,
  choices = rep(NA, 5), ...)
}
\arguments{
\item{GUI}{character: to use a method.}

\item{item}{numeric: index of item.}

\item{question}{character: description of the question.}

\item{choices}{character: description of each choice (indexed as vector).}

\item{templateID}{character: \strong{Concerto} Template ID (or name) to show.}

\item{templateWorkspaceID}{numeric: \strong{Concerto} Workspace ID where the \code{templateID} is located.}

\item{submit}{character: text to display on submit button.}

\item{console.force}{logical: \code{TRUE} to force read a response from console just like from the method \code{console}.}

\item{\dots}{arguments: to pass it on functions.}
}
\value{
{numeric: index of choice by response}
}
\description{
The \code{stepCAT.choice} shows the choices for the question and reads a response from user.
}
\details{
The \code{stepCAT.choice} as generic function has three methods:
\describe{
\item{\code{GUI} of class \code{'console'}}{if argument \code{item} has \code{NULL} value it did not join with argument \code{question} to show;\cr argument \code{choices} must be a vector at least one element or else default value is assumed.}
\item{\code{GUI} of class \code{'concerto'}}{argument \code{choices} must be a positive integer or else default value is assumed;\cr if argument \code{templateID} has \code{NULL} value the \emph{HTML} code is generated and argument \code{templateWorkspaceID} is ignored;\cr argument \code{templateWorkspaceID} must be a positive integer or else \code{NULL} value is assumed;\cr if argument \code{templateWorkspaceID} has \code{NULL} value the \strong{Concerto} session variable \code{concerto$workspaceID} is assumed.}
\item{\code{GUI} of class \code{'tcltk'}}{if argument \code{item} has \code{NULL} value it is did not join with argument \code{question} to show; argument \code{choices} must be a vector at least one element or else default value is assumed.}
}
}
\examples{
## stepCAT.choice example 1

}
\seealso{
\code{\link{stepCAT.by}}, \href{https://code.google.com/p/concerto-platform/}{Concerto Platform}
}

